[~] migrate SteamStorageAPI.SDK to 2.0.0 [~] migrate to .NET 10 [+] add README.md [+] add github workflow

+++ SteamStorage/Models/Tools/UtilityModels/ArchiveGroupModel.cs
@@ -11,7 +11,7 @@ public class ArchiveGroupModel : ExtendedBaseGroupModel
 
     public string SoldSumString { get; }
 
-    public double Change { get; }
+    public decimal Change { get; }
 
     #endregion Properties
 
@@ -25,7 +25,7 @@ public class ArchiveGroupModel : ExtendedBaseGroupModel
         decimal buySum,
         decimal soldSum,
         string currencyMark,
-        double change,
+        decimal change,
         DateTime dateCreation,
         string? description) : base(groupId, title, colour, count, dateCreation, description)
     {
